Fix error handling of httpc_manager and its starter process
This commit fixes the following cases:

    o  If the starter process fails to start httpc_handler,
       httpc_manager doesn't send any answer to the original caller and
       use gen_server:reply/2 incorrectly ("From" is a PID and not what
       was given to handle_call/3). Now, it sends an answer to the caller.
       It doesn't use gen_server:reply/2 anymore; handle_call/3 already
       provided an answer anyway.

    o  If httpc_handler:connect_and_send/2 returns an error,
       both httpc_manager and httpc_handler send an answer to the
       original caller. Now, only httpc_handler does it.

Another problematic case still exists: if the starter process crashes or
is killed (see maybe_handle_terminating_starter/3), httpc_manager can't
determine if httpc_handler was ever started and if the latter will send
an answer. In this commit, httpc_manager sends an answer too. This could
be improved.

While in this function, the same incorrect gen_server:reply/2 as above
is removed.